Match the documentation.
Only return non-nil from gxref-xref-backend if gxref-project-root-dir is set or
if global finds a database. Previously it always returned non-nil, so it
overrode any xref-backend later in xref-backend-functions.

This means you can use the documentation's advice of putting (add-to-list
'xref-backend-functions 'gxref-xref-backend) in your init file without messing
with other backends, for example the emacs lisp backend.